+++ src/affiliations-authors/affiliations-authors.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { SequelizeModule } from '@nestjs/sequelize';
+import { AffiliationAuthor } from 'src/database/models/affiliation-author.model';
+
+@Module({
+  imports: [SequelizeModule.forFeature([AffiliationAuthor])],
+})
+export class AffiliationsAuthorsModule {}