+++ src/database/models/mesh.model.ts
@@ -18,14 +18,6 @@ import { ArticleMesh } from './article-mesh.model';
 })
 export class Mesh extends Model {
   @Column({
-    type: DataType.INTEGER,
-    primaryKey: true,
-    autoIncrement: true,
-  })
-  //@ts-expect-error // Model already defines id
-  id: number;
-
-  @Column({
     type: DataType.STRING,
     allowNull: false,
   })