+++ src/languages/languages.module.ts
@@ -1,8 +1,11 @@
 import { Module } from '@nestjs/common';
 import { SequelizeModule } from '@nestjs/sequelize';
 import { Language } from 'src/database/models/language.model';
+import { LanguagesService } from './languages.service';
 
 @Module({
   imports: [SequelizeModule.forFeature([Language])],
+  providers: [LanguagesService],
+  exports: [LanguagesService],
 })
 export class LanguagesModule {}